[flexbe_testing] Move state copies for testing into behavior
